chore(refactor): simplify cli package consolidate subpackages and remove complexity.
